requirements.txt
yfinance==0.2.36
requests==2.31.0
anthropic==0.18.1
sendgrid==6.11.0
python-telegram-bot==20.8
pyyaml==6.0.1
pytz==2024.1
